renamed SubredditBanner to SectionBanner
